tests: bump patch coverage
