+++ packages/zowe-explorer/__tests__/__unit__/extension.unit.test.ts
@@ -211,8 +211,6 @@ async function createGlobalMocks() {
             "zowe.uss.pasteUssFile",
             "zowe.uss.copyUssFile",
             "zowe.uss.openWithEncoding",
-            "zowe.diff.useLocalContent",
-            "zowe.diff.useRemoteContent",
             "zowe.onJobChanged",
             "zowe.jobs.deleteJob",
             "zowe.jobs.runModifyCommand",
@@ -252,6 +250,8 @@ async function createGlobalMocks() {
             "zowe.editHistory",
             "zowe.promptCredentials",
             "zowe.profileManagement",
+            "zowe.diff.useLocalContent",
+            "zowe.diff.useRemoteContent",
             "zowe.openRecentMember",
             "zowe.searchInAllLoadedItems",
             "zowe.ds.deleteProfile",
