+++ packages/zowe-explorer/src/shared/init.ts
@@ -13,7 +13,7 @@ import * as globals from "../globals";
 import * as vscode from "vscode";
 import * as refreshActions from "./refresh";
 import * as sharedActions from "./actions";
-import { FileManagement, IZoweTree, IZoweTreeNode, Validation } from "@zowe/zowe-explorer-api";
+import { FileManagement, IZoweTree, IZoweTreeNode, Validation, ZoweScheme } from "@zowe/zowe-explorer-api";
 import { ZoweExplorerApiRegister } from "../ZoweExplorerApiRegister";
 import { Profiles } from "../Profiles";
 import { hideTempFolder, moveTempFolder } from "../utils/TempFolder";
@@ -30,6 +30,8 @@ import { ProfileManagement } from "../utils/ProfileManagement";
 import { LocalFileManagement } from "../utils/LocalFileManagement";
 import { TreeProviders } from "./TreeProviders";
 import { IZoweProviders } from "./IZoweProviders";
+import { UssFSProvider } from "../uss/UssFSProvider";
+import { DatasetFSProvider } from "../dataset/DatasetFSProvider";
 
 export function registerRefreshCommand(
     context: vscode.ExtensionContext,
@@ -92,6 +94,25 @@ export function registerCommonCommands(context: vscode.ExtensionContext, provide
     context.subscriptions.push(
         vscode.commands.registerCommand("zowe.profileManagement", async (node: IZoweTreeNode) => {
             await ProfileManagement.manageProfile(node);
+        })
+    );
+
+    context.subscriptions.push(
+        vscode.commands.registerCommand("zowe.diff.useLocalContent", async (localUri) => {
+            if (localUri.scheme === ZoweScheme.USS) {
+                await UssFSProvider.instance.diffOverwrite(localUri);
+            } else if (localUri.scheme === ZoweScheme.DS) {
+                await DatasetFSProvider.instance.diffOverwrite(localUri);
+            }
+        })
+    );
+    context.subscriptions.push(
+        vscode.commands.registerCommand("zowe.diff.useRemoteContent", async (localUri) => {
+            if (localUri.scheme === ZoweScheme.USS) {
+                await UssFSProvider.instance.diffUseRemote(localUri);
+            } else if (localUri.scheme === ZoweScheme.DS) {
+                await DatasetFSProvider.instance.diffUseRemote(localUri);
+            }
         })
     );
